Criado testes de recuperação de dados no banco de dados, criado a persistencia para recuperação dos emprestismos e criado persistencia para o item - parcialmente até agora, precisa ser implementado.

+++ src/main/java/shokunin/group/com/biblioteca/strategy/contracts/EmprestimoStrategy.java
@@ -1,10 +1,12 @@
 package shokunin.group.com.biblioteca.strategy.contracts;
 
+import java.math.BigDecimal;
+
 public interface EmprestimoStrategy {
     int getDiasEmprestimo();
-    double getMultaDiaria();
-    double getMultaMaxima();
-    double getMultaMinima();
+    BigDecimal getMultaDiaria();
+    BigDecimal getMultaMaxima();
+    BigDecimal getMultaMinima();
     int getMaximoItens();
     int getMinimoItens();
     String getTipoUsuario();
